move coerce tests out of base
