+++ tests/Constraints/CoerciveTest.php
@@ -8,9 +8,92 @@
  */
 
 namespace JsonSchema\Tests\Constraints;
+use JsonSchema\Constraints\Constraint;
+use JsonSchema\SchemaStorage;
+use JsonSchema\Uri\UriResolver;
+use JsonSchema\Validator;
 
 class CoerciveTest extends BasicTypesTest
 {
+	/**
+	 * @dataProvider getValidCoerceTests
+	 */
+	public function testValidCoerceCasesUsingAssoc($input, $schema, $checkMode = Constraint::CHECK_MODE_COERCE)
+	{
+		if ($checkMode !== Constraint::CHECK_MODE_COERCE) {
+			$this->markTestSkipped('Test indicates that it is not for "CHECK_MODE_COERCE"');
+		}
+
+		$schema = json_decode($schema);
+		$schemaStorage = new SchemaStorage($this->getUriRetrieverMock($schema), new UriResolver);
+		$schema = $schemaStorage->getSchema('http://www.my-domain.com/schema.json');
+
+		$value = json_decode($input, true);
+		$validator = new Validator($checkMode, $schemaStorage);
+
+		$validator->check($value, $schema);
+		$this->assertTrue($validator->isValid(), print_r($validator->getErrors(), true));
+	}
+
+	/**
+	 * @dataProvider getValidCoerceTests
+	 */
+	public function testValidCoerceCases($input, $schema, $checkMode = Constraint::CHECK_MODE_COERCE)
+	{
+		if ($checkMode !== Constraint::CHECK_MODE_COERCE) {
+			$this->markTestSkipped('Test indicates that it is not for "CHECK_MODE_COERCE"');
+		}
+
+		$schemaStorage = new SchemaStorage($this->getUriRetrieverMock(json_decode($schema)));
+		$schema = $schemaStorage->getSchema('http://www.my-domain.com/schema.json');
+
+		$validator = new Validator($checkMode, $schemaStorage);
+		$validator->check(json_decode($input), $schema);
+
+		$this->assertTrue($validator->isValid(), print_r($validator->getErrors(), true));
+	}
+
+	/**
+	 * @dataProvider getInvalidCoerceTests
+	 */
+	public function testInvalidCoerceCases($input, $schema, $checkMode = Constraint::CHECK_MODE_COERCE, $errors = array())
+	{
+		$checkMode = $checkMode === null ? Constraint::CHECK_MODE_COERCE : $checkMode;
+
+		$schemaStorage = new SchemaStorage($this->getUriRetrieverMock(json_decode($schema)));
+		$schema = $schemaStorage->getSchema('http://www.my-domain.com/schema.json');
+
+		$validator = new Validator($checkMode, $schemaStorage);
+		$validator->check(json_decode($input), $schema);
+
+		if (array() !== $errors) {
+			$this->assertEquals($errors, $validator->getErrors(), print_r($validator->getErrors(),true));
+		}
+		$this->assertFalse($validator->isValid(), print_r($validator->getErrors(), true));
+	}
+
+	/**
+	 * @dataProvider getInvalidCoerceTests
+	 */
+	public function testInvalidCoerceCasesUsingAssoc($input, $schema, $checkMode = Constraint::CHECK_MODE_COERCE, $errors = array())
+	{
+		$checkMode = $checkMode === null ? Constraint::CHECK_MODE_COERCE : $checkMode;
+		if ($checkMode !== Constraint::CHECK_MODE_COERCE) {
+			$this->markTestSkipped('Test indicates that it is not for "CHECK_MODE_COERCE"');
+		}
+
+		$schemaStorage = new SchemaStorage($this->getUriRetrieverMock(json_decode($schema)));
+		$schema = $schemaStorage->getSchema('http://www.my-domain.com/schema.json');
+
+		$validator = new Validator($checkMode, $schemaStorage);
+		$validator->check(json_decode($input, true), $schema);
+
+		if (array() !== $errors) {
+			$this->assertEquals($errors, $validator->getErrors(), print_r($validator->getErrors(), true));
+		}
+		$this->assertFalse($validator->isValid(), print_r($validator->getErrors(), true));
+	}
+
 	public function getValidCoerceTests()
 	{
 		return array(
@@ -54,6 +137,8 @@ class CoerciveTest extends BasicTypesTest
 			)
 		);
 	}
+
+
 
 	public function getInvalidCoerceTests()
 	{
